Update to Node v22 (#127)
This is the Active LTS version of Node.

The previous way `ts-node` was called no longer works in this version of
Node (or v20, the Maintenance LTS version), as it doesn't recognise
files with the `.ts` extension when run. The solution was to switch to
running `node` directly, and registering `ts-node/esm` as a
customisation hook instead. This means it's able to run as part of
Node's module resolution and loading, to transpile TypeScript modules.